watchtower/blob: move code to justice_kit_packet.go
